+++ lib/srtm/cache/file.ex
@@ -10,25 +10,23 @@ defmodule SRTM.Cache.File do
 
   @impl true
   def fetch(path) do
-    if File.exists?(path) do
-      data = File.read!(path)
-      name = Path.basename(path, ".hgt")
-      DataCell.new(name, data)
+    with {:ok, data} <- File.read(path),
+         {:ok, data_cell} <- DataCell.new(Path.basename(path, ".hgt"), data) do
+      {:ok, data_cell}
     else
-      :error
+      {:error, _reason} -> :error
     end
   end
 
   @impl true
   def store(path, data_cell) do
-    data = DataCell.to_binary(data_cell)
-
-    cache_dir = Path.dirname(path)
-    File.mkdir_p!(cache_dir)
-
-    with {:error, reason} <- File.write(path, data) do
-      message = "Writing hgt file '#{path}' failed: #{inspect(reason)}"
-      {:error, %Error{reason: :io_error, message: message}}
+    with :ok <- File.mkdir_p(Path.dirname(path)),
+         :ok <- File.write(path, DataCell.to_binary(data_cell)) do
+      :ok
+    else
+      {:error, reason} ->
+        message = "Caching the HGT file at '#{path}' failed: #{inspect(reason)}"
+        {:error, %Error{reason: :io_error, message: message}}
     end
   end
 end